homework-backend/requirements.txt
Flask==2.1.3
Flask-Cors==3.0.10
Werkzeug==2.2.0
click==8.1.3
colorama==0.4.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
jieba==0.42.1
PyMySQL==1.0.2
bcrypt==4.0.1
PyJWT==2.6.0 